Create dependabot.yml (#180)
Renovate currently doesn't create PRs for vulnerabilities that only exist in transitive dependencies. There's a `transitiveRemediation` flag that can be used, but it only seems to support npm (or doesn't fully work atm).

The alternative would be to add `:maintainLockFilesMonthly` and let Renovate create a PR to update the lockfile monthly (can be set to weekly too), covering any vulnerability but it sounds messy.

This PR adds back dependabot and tell it to only create those vulnerability PRs, complementing renovate.